onlineworld_backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-CORS
Flask-Migrate
python-dotenv
slugify
bleach
requests
reportlab
beautifulsoup4